backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
aws==0.2.5
awscli==1.32.109
bcrypt==4.1.3
boto==2.49.0
boto3==1.34.108
botocore==1.34.109
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
docutils==0.16
email_validator==2.1.1
fabric==3.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
invoke==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
paramiko==3.4.0
passlib==1.7.4
pg8000==1.31.2
prettytable==3.10.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.7.2
s3transfer==0.10.1
scramp==1.4.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
